Cities Generating
Cities now spawn in a variable amount, within the boundaries of the map and not over bodies of water.

TO BE IMPLEMENTED: Cities should not spawn too near each other to create a semblance of borders and degrees of separation between countries. Cities should generate and display names based on the biome they inhabit. Additionally, Cities will spawn smaller towns with similar naming conventions to themselves, exclusively within the same biome in a smaller radius than the city radius.